requirements.txt
arrow==0.14.2
binaryornot==0.4.4
certifi==2019.6.16
chardet==3.0.4
Click==7.0
cookiecutter==1.6.0
defusedxml==0.6.0
Django==2.2.2
django-templated-mail==1.1.1
djangorestframework==3.9.4
djoser==1.7.0
future==0.17.1
idna==2.8
Jinja2==2.10.1
jinja2-time==0.2.0
MarkupSafe==1.1.1
oauthlib==3.0.1
pkg-resources==0.0.0
poyo==0.4.2
PyJWT==1.7.1
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
urllib3==1.25.3
whichcraft==0.5.2
